requirements.txt
pywebview
pynput
